requirements.txt
matplotlib==3.9.4
diffusers==0.32.2
numpy==2.0.2
torch==2.6.0
torchvision==0.21.0
Flask==3.1.0
pillow==11.1.0
gunicorn